pyproject.toml
[project]
name = "academic-investigator"
version = "0.1.0"
description = "Academic investigation toolkit for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "pyalex>=0.14",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0"]

[project.scripts]
acad-inv = "academic_investigator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/academic_investigator"]
